reorganise RenewStatus Cell as both tables use it
